Shader Updated

+++ WinAPIFrameWork/WinGlobal.h
@@ -11,4 +11,4 @@ inline ID3D11DeviceContext* g_context = nullptr;
 inline IDXGISwapChain* g_swapChain = nullptr;
 inline ID3D11RenderTargetView* g_renderTargetView = nullptr;
 
-inline const Vector2 g_viewPort = Vector2(GetSystemMetrics(SM_CXSCREEN), GetSystemMetrics(SM_CYSCREEN));
+inline const Vector2 g_viewPort = Vector2((float)(GetSystemMetrics(SM_CXSCREEN)), (float)(GetSystemMetrics(SM_CYSCREEN)));
